% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_analysis_functions.R
\name{analyze_secular_trends}
\alias{analyze_secular_trends}
\title{Generate Secular Trends Analysis}
\usage{
analyze_secular_trends(enhanced_description, ticker)
}
\arguments{
\item{enhanced_description}{Enhanced business description}

\item{ticker}{Company ticker symbol}
}
\value{
Brief analysis of relevant secular trends
}
\description{
Generate Secular Trends Analysis
}
