% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_consolidated_table.R
\name{create_consolidated_table}
\alias{create_consolidated_table}
\title{Create and Display Formatted GT Table for Financial Metrics}
\usage{
create_consolidated_table(metrics_df)
}
\arguments{
\item{metrics_df}{A tibble containing financial metrics}
}
\value{
A GT table object
}
\description{
Takes a metrics dataframe and creates a formatted GT table
}
