% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_financial_statements.R
\name{fetch_balance_sheet}
\alias{fetch_balance_sheet}
\title{Fetch and Process Balance Sheet Data}
\usage{
fetch_balance_sheet(ticker, api_key, period = "quarterly")
}
\arguments{
\item{ticker}{Character string of the stock ticker symbol}

\item{api_key}{Character string containing the Alpha Vantage API key}

\item{period}{Character string either "annual" or "quarterly"}
}
\value{
A tibble with processed balance sheet data
}
\description{
Fetch and Process Balance Sheet Data
}
