% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_delayed_quote.R
\name{fetch_delayed_quote}
\alias{fetch_delayed_quote}
\title{Make a request to the Alpha Vantage API for delayed quotes}
\usage{
fetch_delayed_quote(ticker, api_key)
}
\arguments{
\item{ticker}{Character string of the stock ticker symbol}

\item{api_key}{Character string containing the Alpha Vantage API key}
}
\value{
List containing the price and timestamp
}
\description{
Make a request to the Alpha Vantage API for delayed quotes
}
