% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_financial_statements.R
\name{fetch_income_statement}
\alias{fetch_income_statement}
\title{Fetch and Process Income Statement Data}
\usage{
fetch_income_statement(ticker, api_key, period = "quarterly")
}
\arguments{
\item{ticker}{Character string of the stock ticker symbol}

\item{api_key}{Character string containing the Alpha Vantage API key}

\item{period}{Character string either "annual" or "quarterly"}
}
\value{
A tibble with processed income statement data
}
\description{
Fetch and Process Income Statement Data
}
