% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_latest_price.R
\name{fetch_latest_price}
\alias{fetch_latest_price}
\title{Make a request to the Alpha Vantage API for latest intraday price}
\usage{
fetch_latest_price(ticker, api_key)
}
\arguments{
\item{ticker}{Character string of the stock ticker symbol}

\item{api_key}{Character string containing the Alpha Vantage API key}
}
\value{
List containing the latest price data
}
\description{
Make a request to the Alpha Vantage API for latest intraday price
}
