% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_split_history.R
\name{fetch_split_history}
\alias{fetch_split_history}
\title{Fetch Stock Split History}
\usage{
fetch_split_history(ticker, api_key)
}
\arguments{
\item{ticker}{Character string of the stock ticker symbol}

\item{api_key}{Character string containing the Alpha Vantage API key}
}
\value{
A dataframe of split events with dates and factors
}
\description{
Fetch Stock Split History
}
