% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_alpha_vantage_request.R
\name{make_alpha_vantage_request}
\alias{make_alpha_vantage_request}
\title{Make a request to the Alpha Vantage API}
\usage{
make_alpha_vantage_request(function_name, ticker, api_key, ...)
}
\arguments{
\item{function_name}{Character string specifying the Alpha Vantage API function to call}

\item{ticker}{Character string of the stock ticker symbol}

\item{api_key}{Character string containing the Alpha Vantage API key}

\item{...}{Additional parameters to pass to the API request}
}
\value{
List containing the JSON response from the API
}
\description{
Sends a request to the Alpha Vantage API with the specified function name, ticker and API key.
Handles errors and returns the JSON response body.
}
